requirements.txt
discord.py[voice]==2.3.2
yt-dlp
aiohttp
PyNaCl
python-dotenv
